add remove_duplicates

+++ search-server/request_queue.cpp
@@ -9,13 +9,13 @@ RequestQueue::RequestQueue(const SearchServer& search_server)
 //-------------------------------------------------------------------------------------------------------------
 std::vector<Document> RequestQueue::AddFindRequest(const std::string& raw_query, DocumentStatus status) {
     const auto result = search_server_.FindTopDocuments(raw_query, status);
-    AddRequest(result.size());
+    AddRequest(static_cast<int>(result.size()));
     return result;
 }
 //-------------------------------------------------------------------------------------------------------------
 std::vector<Document> RequestQueue::AddFindRequest(const std::string& raw_query) {
     const auto result = search_server_.FindTopDocuments(raw_query);
-    AddRequest(result.size());
+    AddRequest(static_cast<int>(result.size()));
     return result;
 }
 //-------------------------------------------------------------------------------------------------------------
